inventory-service/requirements.txt
fastapi>=0.100,<1.0
uvicorn[standard]>=0.23,<1.0
pydantic>=2.8,<3.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
grpcio>=1.60,<2.0
grpcio-tools>=1.60,<2.0
grpcio-health-checking>=1.60,<2.0
pika>=1.3.2,<2.0
